Update audio-context-fix to work for in Chrome.
